+++ apps/regression/graph_controller.cpp
@@ -433,7 +433,7 @@ void GraphController::setRoundCrossCursorView() {
   bool round = *m_selectedDotIndex < 0;
   if (round) {
     // Set the color although the cursor view stays round
-    assert(*m_selectedSeriesIndex < Palette::numberOfDataColors());
+    assert(*m_selectedSeriesIndex < static_cast<int>(Palette::numberOfDataColors()));
     m_roundCursorView.setColor(Palette::DataColor[*m_selectedSeriesIndex]);
   }
   CursorView * nextCursorView = round ? static_cast<Shared::CursorView *>(&m_roundCursorView) : static_cast<Shared::CursorView *>(&m_crossCursorView);
